fix(attention): use int64 for paged KV-cache addressing in MHA kernels
The MHA prefill/decode Triton kernels computed the KV-cache element offset
as `physical_page * PAGE_SIZE + page_offsets` in int32. For large KV caches
(e.g. Qwen3.5 with millions of cache tokens) `physical_page * PAGE_SIZE *
stride` exceeds the int32 range and silently wraps to a negative address,
so the load returns zeros and full-attention output collapses to zero --
producing catastrophic numeric degradation while linear-attention layers
still look healthy. Promote the page index to int64 before the offset
computation in mha_prefill and mha_decode (both the grouped and ungrouped
stage-1 paths).

Also add a TOKENSPEED_DISABLE_ALLREDUCE_FUSION escape hatch in
resolve_communication: the auto-enabled fused allreduce+RMSNorm path routes
through the Iris symmetric-heap collective, whose fd-passing setup can
deadlock on some TP topologies; this flag lets operators fall back to the
unfused allreduce path.

+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/triton/mha_prefill.py
@@ -170,7 +170,10 @@ def _fwd_kernel(
                     mask=mask_n,
                     other=0,
                 )
-                offs_kv_loc = physical_pages * PAGE_SIZE + page_offsets
+                # Promote to int64: for large KV caches the element offset
+                # (physical_page * page_size * stride) can exceed the int32 range
+                # and silently wrap to a negative address, yielding zero output.
+                offs_kv_loc = physical_pages.to(tl.int64) * PAGE_SIZE + page_offsets
                 offs_k = (
                     offs_kv_loc[None, :] * stride_buf_kbs
                     + cur_kv_head * stride_buf_kh
